refactor: 결제 서비스 코드 정리 및 버그 수정(PDSH-1)
