+++ payment/src/main/java/com/live_commerce/payment/presentation/controller/PaymentControllerV2.java
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.live_commerce.payment.application.dto.request.PaymentApproveRequestDto;
 import com.live_commerce.payment.application.dto.request.PaymentReadyRequestDto;
-import com.live_commerce.payment.application.dto.request.PaymentRefundResponseDto;
+import com.live_commerce.payment.application.dto.response.PaymentRefundResponseDto;
 import com.live_commerce.payment.application.dto.request.PaymentSearchCondition;
 import com.live_commerce.payment.application.dto.response.PaymentApproveResponseDto;
 import com.live_commerce.payment.application.dto.response.PaymentGetResponseDto;
